sysadmin scripts, control panel bug

+++ sysadmin/update.sh
@@ -12,8 +12,7 @@ cd /var/www/gamocosm
 git checkout release
 git pull origin release
 
-rvm use 2.2
-bundle install
+./run.sh bundle install
 RAILS_ENV=production ./run.sh --bundler rake assets:precompile
 RAILS_ENV=production ./run.sh --bundler rake db:migrate
 RAILS_ENV=test ./run.sh rake db:migrate
@@ -21,8 +20,8 @@ RAILS_ENV=test ./run.sh rake db:migrate
 
 TIMESTAMP="$(date +'%Y_%m_%d-%H:%M')"
 mkdir -p log/archive
-mv log/production.log "log/archive/production.$TIMESTAMP.log"
-mv log/sidekiq.log "log/archive/sidekiq.$TIMESTAMP.log"
+[ -e log/production.log ] && mv log/production.log "log/archive/production.$TIMESTAMP.log"
+[ -e log/sidekiq.log ] && mv log/sidekiq.log "log/archive/sidekiq.$TIMESTAMP.log"
 
 touch tmp/restart.txt
